Finished hw for the day!
Again, I'm leaving out the vast majority of my work, just in case putting anything up counts as a violation of the honor code.
In 103, we had to practice using this typescipting tool called LaTech that's used for mathematical writing. It'll take a lot fo getting used to, but it seems really helpful.
We have to learn how to regularly test in 106b, which is something I have to get used to. They're teaching us how to make our own valid tests (using a specific Stanford tool tho.)